Modified some things. See Description
Merged the LCD and LED tasks into controller. This will reduce the
number of threads running.

made speaker.c a little more stable when there is a problem reading the
SD card.

DEBUG2 is defined currently and there are some tasks commented out, so
put those back in before testing.